Add post controller

+++ src/main/java/com/hungtin/tako/takobackend/post/service/PostService.java
@@ -1,5 +1,6 @@
 package com.hungtin.tako.takobackend.post.service;
 
+import com.hungtin.tako.takobackend.auth.model.UserAccount;
 import com.hungtin.tako.takobackend.auth.service.AuthService;
 import com.hungtin.tako.takobackend.post.http.CreatePostRequest;
 import com.hungtin.tako.takobackend.post.http.PostResponse;
@@ -7,7 +8,9 @@ import com.hungtin.tako.takobackend.post.http.UpdatePostRequest;
 import com.hungtin.tako.takobackend.post.http.mapping.PostMapper;
 import com.hungtin.tako.takobackend.post.model.Post;
 import com.hungtin.tako.takobackend.post.repo.PostRepo;
+import com.hungtin.tako.takobackend.user.User;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -27,10 +30,19 @@ public class PostService {
   public PostResponse create(CreatePostRequest request) {
     Post post = postMapper.transform(request);
     post.setUser(authService.getCurrentUser().getUser());
+
+    post = postRepo.save(post);
+
     return postMapper.transform(post);
   }
 
+  @Transactional
   public void delete(Long id) {
+    // check if the current user is the author of the post
+
+    if (isCurrentUserOwnPost(id)) {
+      throw new IllegalArgumentException("Post id is not valid for delete by current user: " + id);
+    }
     postRepo.deleteById(id);
   }
 
@@ -41,7 +53,24 @@ public class PostService {
 
   @Transactional
   public PostResponse update(UpdatePostRequest request) {
+    if (isCurrentUserOwnPost(request.getId())) {
+      throw new IllegalArgumentException(
+          "Post id is not valid for delete by current user: " + request.getId());
+    }
+
     Post post = postMapper.transform(request);
     return postMapper.transform(postRepo.save(post));
+  }
+
+
+  @Transactional(readOnly = true)
+  public boolean isCurrentUserOwnPost(Long postId) {
+    UserAccount userAccount = authService.getCurrentUser();
+    Optional<Long> optUserId = postRepo.findById(postId)
+        .map(Post::getUser)
+        .map(User::getId);
+
+    return optUserId.isEmpty()
+        || !optUserId.get().equals(userAccount.getUser().getId());
   }
 }
